Reworked the texture system so that we now have a texture interface that's abstract, and we have a opengl concrete texture implementation so that we don't expose any opengl code to the core of the engine.
Also refactored the texture loader and texture manager to separate concerns between both, and also the assimp loader as some of the work for texture extraction needed to be delegated back to the assimp importer.

Created assimp facade too so that we dont expose everything to the engine, can be easily modified and extended and prevents changes to assimp concrete class from having breaking changes.

Also created a helper function to extract raw texture data from an embedded file and pass it along to a texture class.